Fine tuning the configuration.
git-svn-id: http://ikube.googlecode.com/svn/trunk@981 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/cluster/hzc/IndexContextListener.java
@@ -47,6 +47,9 @@ public class IndexContextListener implements MessageListener<Object> {
 			dataBase.persist(remoteIndexContext);
 		} else {
 			// Check the time stamp of the contexts
+			if (remoteIndexContext.getTimestamp() == null || localIndexContext.getTimestamp() == null) {
+				return;
+			}
 			if (remoteIndexContext.getTimestamp().after(localIndexContext.getTimestamp())) {
 				LOGGER.info("Adding index context : " + remoteIndexContext);
 				dataBase.remove(localIndexContext);
